+++ opendrift/models/oceandrift.py
@@ -121,8 +121,9 @@ class OceanDrift(OpenDriftSimulation):
             'drift:tabularised_stokes_drift_fetch': {'type': 'enum', 'enum': ['5000', '25000', '50000'], 'default': '25000',
                 'level': self.CONFIG_LEVEL_ADVANCED, 'description':
                 'The fetch length when using tabularised Stokes drift.'},
-            'drift:lift_to_seafloor': {'type': 'bool', 'default': True,
-                'description': 'If True, elements hitting/penetrating seafloor, are lifted to seafloor height. The alternative (False) is to deactivate elements).',
+            'general:seafloor_action': {'type': 'enum', 'default': 'lift_to_seafloor',
+                'enum': ['none', 'lift_to_seafloor', 'deactivate', 'previous'],
+                'description': '"deactivate": elements are deactivated; "lift_to_seafloor": elements are lifted to seafloor level; "previous": elements are moved back to previous position; "none"; seafloor is ignored.',
                 'level': self.CONFIG_LEVEL_ADVANCED},
             'drift:truncate_ocean_model_below_m': {'type': 'float', 'default': None,
                 'min': 0, 'max': 10000, 'units': 'm',
@@ -212,7 +213,7 @@ class OceanDrift(OpenDriftSimulation):
         bottom = np.where(self.elements.z < Zmin)
         if len(bottom[0]) > 0:
             logger.debug('%s elements reached seafloor, set to bottom' % len(bottom[0]))
-            self.elements.z[bottom] = Zmin[bottom]
+            self.interact_with_seafloor()
             self.bottom_interaction(Zmin)
 
     def surface_stick(self):
@@ -413,7 +414,7 @@ class OceanDrift(OpenDriftSimulation):
             bottom = np.where(self.elements.z < Zmin)
             if len(bottom[0]) > 0:
                 logger.debug('%s elements reached seafloor, set to bottom' % len(bottom[0]))
-                self.elements.z[bottom] = Zmin[bottom]
+                self.interact_with_seafloor()
                 self.bottom_interaction(Zmin)
 
             if store_depths is not False: