+++ opendrift/readers/basereader/variables.py
@@ -15,7 +15,6 @@ class ReaderDomain(Timeable):
     Projection, spatial and temporal domain of reader.
     """
     name = None
-    simulation_SRS = False
 
     proj4 = None
     proj = None
@@ -393,12 +392,16 @@ class Variables(ReaderDomain):
 
     environment_mappers = []
     environment_mappings = {
-        'wind_from_speed_and_direction': {
+        'wind_from_speed_and_direction_from': {
+            'input': ['wind_speed', 'wind_from_direction'],
+            'output': ['x_wind', 'y_wind'],
+            'method':
+            lambda reader, var: reader.wind_from_speed_and_direction(var)},
+        'wind_from_speed_and_direction_to': {
             'input': ['wind_speed', 'wind_to_direction'],
             'output': ['x_wind', 'y_wind'],
             'method':
-            lambda reader, var: reader.wind_from_speed_and_direction(var)
-        },
+            lambda reader, var: reader.wind_from_speed_and_direction(var)},
         'testvar': {
             'input': ['sea_ice_thickness'],
             'output': ['istjukkleik']
@@ -429,10 +432,12 @@ class Variables(ReaderDomain):
                          (self.buffer, self.name, max_speed))
 
     def wind_from_speed_and_direction(self, env):
-        north_wind = env['wind_speed'] * np.cos(
-            np.radians(env['wind_to_direction']))
-        east_wind = env['wind_speed'] * np.sin(
-            np.radians(env['wind_to_direction']))
+        if 'wind_from_direction' in env:
+            wfd = env['wind_from_direction']
+        else:
+            wfd = -env['wind_to_direction']
+        north_wind = -env['wind_speed']*np.cos(np.radians(wfd))
+        east_wind = -env['wind_speed']*np.sin(np.radians(wfd))
         env['x_wind'] = east_wind
         env['y_wind'] = north_wind
         # Rotating might be necessary generally
@@ -620,8 +625,8 @@ class Variables(ReaderDomain):
 
         # Rotating vectors fields
         if rotate_to_proj is not None:
-            if self.simulation_SRS is True:
-                logger.debug('Reader SRS is the same as calculation SRS - '
+            if self.proj.crs.is_geographic:
+                logger.debug('Reader projection is latlon - '
                              'rotation of vectors is not needed.')
             else:
                 vector_pairs = []